+++ OpenRA.Mods.Common/Pathfinder/PathSearch.cs
@@ -15,7 +15,6 @@ using System.Linq;
 using System.Runtime.CompilerServices;
 using OpenRA.Mods.Common.Traits;
 using OpenRA.Primitives;
-using OpenRA.Traits;
 
 namespace OpenRA.Mods.Common.Pathfinder
 {
